requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.3
Django==4.2.20
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
reportlab==4.4.3
sqlparse==0.5.3
typing_extensions==4.13.1
